Moved examples for widgets from examples/ to widget/example.py and fixed imports
