+++ docs/widgets/hiding/index.rst
@@ -4,11 +4,11 @@ Hiding widgets
 Actually every widget can be hidden including texts, buttons, groups and so on.
 It is managed by ``when`` attribute. It can be either a data key, a predicate function or a F-filter (from ``magic-filter``)
 
-.. literalinclude:: examples/widgets/whenable.py
+.. literalinclude:: ./example.py
 
-.. image::  resources/whenable.png
+.. image::  /resources/whenable.png
 
 If you only change data setting ``"extended": True`` the window will look differently
 
 
-.. image::  resources/whenable_extended.png
+.. image::  /resources/whenable_extended.png
