+++ docs/widgets/keyboard/column/index.rst
@@ -4,6 +4,6 @@ Column
 
 **Column** widget is like a row, but places everything in a column, also ignoring hierarchy.
 
-.. literalinclude:: examples/widgets/column.py
+.. literalinclude:: ./example.py
 
-.. image:: resources/column.png
+.. image:: /resources/column.png
