+++ docs/widgets/keyboard/group/example_width.py
