+++ docs/widgets/keyboard/row/index.rst
@@ -5,6 +5,6 @@ Row
 **Row** widget is used to place all buttons inside single row.
 You can place any keyboard widgets inside it (for example buttons or groups) and it will ignore any hierarchy and just place telegram buttons in a row.
 
-.. literalinclude:: examples/widgets/row.py
+.. literalinclude:: ./example.py
 
-.. image:: resources/row.png
+.. image:: /resources/row.png
