+++ docs/widgets/keyboard/scrolling_group/index.rst
@@ -5,7 +5,7 @@ ScrollingGroup
 **ScrollingGroup** widget combines buttons into pages with the ability to scroll forward and backward and go to the last or first page with buttons.
 You can set the ``height`` and ``width`` of the keyboard. If there are not enough buttons for the last page, the keyboard will be filled with empty buttons keeping the specified height and width.
 
-.. literalinclude:: examples/widgets/scrolling_group.py
+.. literalinclude:: ./example.py
 
-.. image:: resources/scrolling_group1.png
-.. image:: resources/scrolling_group2.png
+.. image:: /resources/scrolling_group1.png
+.. image:: /resources/scrolling_group2.png
