+++ docs/widgets/keyboard/url/example.py
@@ -1,0 +1,7 @@
+from aiogram_dialog.widgets.kbd import Url
+from aiogram_dialog.widgets.text import Const
+
+go_btn = Url(
+    Const("Github"),
+    Const('https://github.com/Tishka17/aiogram_dialog/'),
+)
