+++ docs/widgets/text/case/index.rst
@@ -5,4 +5,4 @@ Case
 To select one of the texts depending on some condition you should use ``Case``.
 The condition can be either a data key or a function:
 
-.. literalinclude:: examples/widgets/case.py
+.. literalinclude:: ./example.py
