+++ docs/widgets/text/multi/index.rst
@@ -4,4 +4,4 @@ Multi
 
 To combine multiple texts you can use ``Multi`` widget. You can use any texts inside it. Also you can provide a string separator. In simple cases you can just concatenate widgets using ``+`` operator.
 
-.. literalinclude:: examples/widgets/multi.py
+.. literalinclude:: ./example.py
